@@ -124,7 +124,11 @@ public class Version implements Comparable<Version>, ToXContentFragment {
     public static final Version V_6_7_0 = new Version(V_6_7_0_ID, org.apache.lucene.util.Version.LUCENE_7_7_0);
     public static final int V_7_0_0_ID = 7000099;
     public static final Version V_7_0_0 = new Version(V_7_0_0_ID, org.apache.lucene.util.Version.LUCENE_8_0_0);
-    public static final Version CURRENT = V_7_0_0;
+    public static final int V_7_1_0_ID = 7010099;
+    public static final Version V_7_1_0 = new Version(V_7_1_0_ID, org.apache.lucene.util.Version.LUCENE_8_0_0);
+    public static final int V_8_0_0_ID = 8000099;
+    public static final Version V_8_0_0 = new Version(V_8_0_0_ID, org.apache.lucene.util.Version.LUCENE_8_0_0);
+    public static final Version CURRENT = V_8_0_0;
 
 
     static {
@@ -138,6 +142,10 @@ public class Version implements Comparable<Version>, ToXContentFragment {
 
     public static Version fromId(int id) {
         switch (id) {
+            case V_8_0_0_ID:
+                return V_8_0_0;
+            case V_7_1_0_ID:
+                return V_7_1_0;
             case V_7_0_0_ID:
                 return V_7_0_0;
             case V_6_7_0_ID:
