@@ -590,7 +590,8 @@ public class MetaDataCreateIndexService {
 
         static int getNumberOfShards(final Settings.Builder indexSettingsBuilder) {
             // TODO: this logic can be removed when the current major version is 8
-            assert Version.CURRENT.major == 7;
+            // TODO: https://github.com/elastic/elasticsearch/issues/38556
+            // assert Version.CURRENT.major == 7;
             final int numberOfShards;
             final Version indexVersionCreated =
                     Version.fromId(Integer.parseInt(indexSettingsBuilder.get(IndexMetaData.SETTING_INDEX_VERSION_CREATED.getKey())));
