@@ -62,14 +62,14 @@ import static java.util.Collections.unmodifiableMap;
  */
 public abstract class FileRestoreContext {
 
-    private static final Logger logger = LogManager.getLogger(FileRestoreContext.class);
+    protected static final Logger logger = LogManager.getLogger(FileRestoreContext.class);
 
-    private final String repositoryName;
-    private final IndexShard indexShard;
-    private final RecoveryState recoveryState;
-    private final SnapshotId snapshotId;
-    private final ShardId shardId;
-    private final int bufferSize;
+    protected final String repositoryName;
+    protected final IndexShard indexShard;
+    protected final RecoveryState recoveryState;
+    protected final SnapshotId snapshotId;
+    protected final ShardId shardId;
+    protected final int bufferSize;
 
     /**
      * Constructs new restore context
@@ -183,7 +183,6 @@ public abstract class FileRestoreContext {
                 // list of all existing store files
                 final List<String> deleteIfExistFiles = Arrays.asList(store.directory().listAll());
 
-                // restore the files from the snapshot to the Lucene store
                 for (final BlobStoreIndexShardSnapshot.FileInfo fileToRecover : filesToRecover) {
                     // if a file with a same physical name already exist in the store we need to delete it
                     // before restoring it from the snapshot. We could be lenient and try to reuse the existing
@@ -196,10 +195,9 @@ public abstract class FileRestoreContext {
                         logger.trace("[{}] [{}] deleting pre-existing file [{}]", shardId, snapshotId, physicalName);
                         store.directory().deleteFile(physicalName);
                     }
-
-                    logger.trace("[{}] [{}] restoring file [{}]", shardId, snapshotId, fileToRecover.name());
-                    restoreFile(fileToRecover, store);
                 }
+
+                restoreFiles(filesToRecover, store);
             } catch (IOException ex) {
                 throw new IndexShardRestoreFailedException(shardId, "Failed to recover index", ex);
             }
@@ -231,6 +229,14 @@ public abstract class FileRestoreContext {
             }
         } finally {
             store.decRef();
+        }
+    }
+
+    protected void restoreFiles(List<BlobStoreIndexShardSnapshot.FileInfo> filesToRecover, Store store) throws IOException {
+        // restore the files from the snapshot to the Lucene store
+        for (final BlobStoreIndexShardSnapshot.FileInfo fileToRecover : filesToRecover) {
+            logger.trace("[{}] [{}] restoring file [{}]", shardId, snapshotId, fileToRecover.name());
+            restoreFile(fileToRecover, store);
         }
     }
 
